requirements.txt
sortedcontainers
pytest
requests
PyHamcrest
PyYAML
allure-pytest
protobuf
googleapis-common-protos
grpcio
grpcio-tools
moviepy
emoji
dialog_api